refactor: remove unused demo code
