--- apps/demo/app/Framework.ts
@@ -1,2 +1,0 @@
-export const validFrameworks = ["antd", "material-ui", "custom"];
-export type Framework = "antd" | "material-ui" | "custom";
