+++ apps/demo/config/blocks/Hero/quotes.ts
