+++ apps/demo/config/blocks/Stats/index.tsx
